+++ src/pages/createBill/addExpenseStep/index.tsx
@@ -1,16 +1,15 @@
 import { useLayoutEffect, useMemo, useRef, useState } from 'react';
 import { FormProvider, useFieldArray, useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
-import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { Close } from '@/assets/svgs/icon';
 import Header from '@/common/components/Header';
 import { BaseFunnelStepComponentProps } from '@/common/types/useFunnel.type';
 import { Group } from '@/common/types/group.type';
+import useCreateExpense from '@/common/queries/expense/useCreateExpense';
 import {
   Expense,
   ExpenseFormSchema,
 } from '@/pages/createBill/types/expense.type';
-import expense from '@/service/apis/expense';
 import group from '@/service/apis/group';
 import BillFormCard from './components/FormCard';
 import getTotalExpense from '../utils/getTotalExpense';
@@ -30,16 +29,7 @@ interface AddExpenseStepProps
 function AddExpenseStep({ moveToNextStep }: AddExpenseStepProps) {
   const lastFormCardRef = useRef<HTMLDivElement | null>(null);
   const [groupInfo, setGroupInfo] = useState<Group | null>(null);
-  const queryClient = useQueryClient();
-  const mutation = useMutation({
-    mutationFn: expense.create,
-    onSuccess: () => {
-      queryClient.invalidateQueries({
-        queryKey: ['expenses'], // TODO : GroupToken 추가 필요함
-      });
-      moveToNextStep?.();
-    },
-  });
+  const mutation = useCreateExpense({ moveToNextStep });
   const formMethods = useForm({
     resolver: zodResolver(ExpenseFormSchema),
     mode: 'onChange', // 폼들의 필수 입력값이 모두 입력되었을 때 '다음' 버튼을 활성화시키기 위함
